Fix: Can't call method "_read" on an undefined value
Thank you very much for you module, which we use on binary.com

On production servers we get multiple errors like:

EventLoop error: I/O watcher failed: Can't call method "_read" on an undefined value at ../local/lib/perl5/Mojo/Redis2.pm line 474.

We wasn't able to catch/reproduce the issue locally/isolated, but with the proposed patch the errors gone. So, we decided to send the patch to upstream

I hope it is rather safe to apply the pathc.